Add /info to retrieve public key
